flask_app/requirements.txt
openai
pymupdf
pandas
python-docx
flask
requests
pdf2image
pillow
werkzeug

